fetcher/requirements.txt
requests==2.32.3
pymongo==4.7.3
python-dateutil==2.9.0.post0
pytz==2024.1
